first pass at converting to swagger 2.0
This commit converts most of the current code to the swagger 2.0 spec.

JSONSubTypes are not yet supported.
ExtraApiDeclarations are not yet supported.
Security Definitions are not yet supported.
Most new features in swagger 2.0 are not yet supported.

Tests need to be updated to swagger 2.0.
README.md needs to be updated.